requirements.txt
fire==0.4.0
gnureadline==8.1.2
networkx==2.6.3
numpy==1.21.6
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
Pillow==9.1.1
six==1.16.0
termcolor==1.1.0
tqdm==4.64.0
